Fix minor FAQ bug where answer won't close
It was working whereby if a new question was clicked the old one would close. But if an open question was clicked a second time, it would not close. This fixes that issue.
Also updated the index variable name and added it as a key to stop Node from complaining.