Refactor button drop-shadow shade
Rename from bgShade to haloShade and invert the value passed in. Much more sensible to target the shade we want rather than inferring it based on background. Sometimes (as in the case of the careers page here) we want light on light.
Also makes the string interpolation for the Button component more elegant.