Install encoding package to resolve warning
This was flagged as an unresolved dependency by Node in the shell when the 'weight' screening quiz page had loaded. It occurs because Firebase is being imported to a file in the app directory.
This is probably a temporary issue with Next and encoding can be uninstalled when it's resolved.