Log message if user account already created
Just log a message to console if account has already been created for the email in the screening quiz data being saved. Might have a toast pop up later.
We have previously also covered the case where the email matches but no account where we then overwrite the data. The only case without remark is when screening data is saved that's not associated with any existing email in Firestore.