Move homepage into subdirectory
Move the Home page into the home directory so that it's at the same level of nesting as the other pages of the main site. Then replace the root component (i.e. at path '/') with a redirect.
This is essentially a refactor for now as existing functionality is maintained. However, it will allow the removal of the default layout and replacement with one that covers just the main site (but does the same thing).