Nest all non-quiz pages in main directory
Create a new directory called main. Then move all website pages apart from the ones that comprise the screening quiz to it. This will facilitate the application of a layout to these only and not to the screening route.
Updated the redirect path in the root entry page and the background video src on the homepage so that everything still renders. Layout links are broken though.