Apply navbar offset to main layout only
Rather than applying the top margin to offset the page body everywhere, localise it to just the main pages. This is because it's not required on the quiz pages as we're never hiding content behind the navbar there.
Also fixes min-height issue where the quiz pages were scrollable due to little content. This could have been fixed without this, but it would have been much less elegant.