Tidy up logic on weight page of screening quiz
No need to check if email exists here any longer. It always exists as would have been saved on the email page.
Can also remove some of the checks when saving screening data. Don't need to check if account is already created as this was done on the email page too.