Check data types of screening fields in Firestore rules
Just checking types are correct for the non-string fields of the screening quiz data before saving to Firestore.
This acts as a further restriction on write actions to the users collection. Reduces the probability that a malicious user could write to the database programatically without using the website form pages.