Fix bug where BMI doesn't show on results page
The restructuring of the weight page caused the figure for the BMI on the results page not to display. This was because it was updating immediately and no longer required the useFreshCookie hook. The cookie can just be used directly as usual.
The hook didn't work as there was no fresh and stale cookies to compare (just a fresh one). The setInterval continued indefinitely due to the conditional never being entered that terminates the function.