Nest screening quiz pages in signup directory
Best to nest in a new signup directory as it will share a layout with quite a few other categories of pages. Want to avoid duplicating that layout file which is now the only one not nested in signup.
Also updated the link URLs for the buttons on each page.