Disable create-account button based on helperText
Need to disable the button that creates a new account based on whether helper text is displayed. This is what tracks the error state of the password field.
Moved helperText and it's setter to the parent (create-account page) so it could be shared with both the Password component and the button.