use setQuizCookie with saveScreeningData
Better to set the cookie first and then save the screening data all in the one function. Much clearer what is happening that way.
Previous pattern of wrapping the standardised quiz button in a div to achieve the composite functionality was messy. We no longer want the navigation that entailed.