Redirect to payments page on successful login
Navigate to the payments page after a successful login so that something happens.
When a user is fully signed up, we'll need to redirect to the patient dashboard. However, when signup is only partially complete we'll need to track at which point in the process the user has reached before and navigate there instead.