Add optional delay to useCookieState hook
We don't usually want a delay, so it can default to zero. However, there are cases where we require one.
An example is where the initial state (i.e. the setter of which is passed in) needs to be calculated on the parent page. Without the delay, the cookie will set first and then the initial state will overwrite it. We want to ensure the cookie loses the race condition and sets after state initialisation.