Increase reCAPTCHA token expiry date
Increased from one week to three months. There's no need to expire so soon as it's not sensitive info - just a bot screen.
Want the increase as there could potentially be a lot of time between when a user starts the screening quiz and when they complete the medical one.