Save email to cookie when token is saved
Save the email to a cookie for up to 3 months which is the same duration as the reCAPTCHA token. This is because the email is needed find the docId and hence the other user data from Firestore. This may be done when saving the medical data which could be a good while later.