Use firebase-admin to save data to Firestore
Using the Admin SDK to save the medical data to Firestore from the server.
Could also have passed the formatted medical data back to the client and saved from there. This would have been possible in the standard way as the useAuth hook exposes the user object here, allowing it to meet the Firestore rules requirement.