Initialise address as empty
This avoids the undesired situation whereby a user first enters an address using all 4 fields. Then they leave and come back, re-entering a different address using only the 3 required fields. The would result in the fourth optional field not being overwritten, giving an incorrect address.
This fix means that field will reset to an empty string if the address is resubmitted per the above.