Get subscription expiry date in webhook
Get the customer ID from the invoice and use to find the customers subscription. Then find when the next payment is due from that.
Hardcoding an existing customer ID associated with a full customer for testing purposes. The reason is otherwise the CLI will create a new customer each time without any subscription details attached.