Disable verification button when loading
Add an isLoading flag that can be turned on when the verification request is being made. This is to stop the user from sending more than one request at a time as the button will be disabled. It also provides the user with a visual clue that something is happening.