Access correct subscription on payments webhook
Best to access to last item in the array as this will be the latest. Only had one item in any testing I did.
It's possible that a new subscription object gets added each month (rather than only at the start of a new type of subscription for a different product). This fix handles that possibility.