Tweak calendar styles
If we have full (i.e. 100%) height as before, there's an area of deadspace at the bottom of the screen. This is due to the Calendly embed element overflowing it's container.
There's no way of modifying this underlying embed directly (it's not exposed) so we just modify the height of the element to avoid deadspace.