Add payment method ID for CLI testing purposes
Create this new constant that can be used during testing of the webhook via the CLI. Use it to replace the value taken from the payment intent (which will be a new value and not the one associated with the subscription buy default when using the CLI).
Also update the other IDs for CLI testing in constants.js.