Use isDev rather than isCli in constants.js
It was premature to get rid of isDev as it's a better flag when selecting the webhook secrets. A variable isLocal would be best, but that's not possible to determine from the node environment.
Still need isCli elsewhere though (i.e. for setting the 3 constants declared at the top of the file).