Track paused subscription on patient dashboard
Get the isPaused status of the subscription from Firestore and save to state. Pass setter to modal so it can be changed when the button there is clicked.
Button gets disabled when clicked so that only one month can be paused at a time. Will get unpaused on next payment.