Update UI with new sub expiry date on dashboard
Pass the new expiry date from the server back to the client. Then set the expiryDate on the main patient dashboard page via the pause-modal.
Before doing all of this, the UI showed the correct info initially but would be incorrect when the page refreshed (i.e. displays updated Firestore data).