Prevent errors on patient dashboard if signup incomplete
Add optional chaining for address1 so that we don't attempt to get that property of undefined if not present.
Also return early from the payments route that fetches card details if not present. Prevents an error on the server.
All of this prevents errors when a user creates an account but does not complete other steps before trying to access the patient dashboard.