Redirect on logout from doctor dashboard
Create an isPageLoaded flag that sets after one second as with patient dashboard. Used to ensure that state has had time to load before checking if redirect is necessary.
Now we can redirect after logout and when URL is navigated to with incorrect or no user. This did not happen before.