Redirect to login if user not pharm user
If the page has loaded but there is no user or they have a non-pharm role, then redirect to login page.
Although it might seem like the user check is redundant, that is not the case. If an existing pharm user logs out, the role will still be saved in state but we want to force the redirect, so listen for the user.