Remove underline from Reformr logo on navbar
There was a thin cyan line under the Reformr logo when the user was on the homepage to show that it was active.
Remove this, but still keep the change in opacity when hovering the logo to show that it's clickable. Effect only takes place when not on homepage as before.