Fix bug on Safari mobile where top of navbar is cropped
When the nav menu was opened on Safari, the Reformr logo and X got obscured by the overlay getting positioned over them.
The fixes that issue, but is not perfect as the user can still scroll the background content. A flicker of this is visible at the bottom of the overlay as the browser navbar collapses or expands on all browsers.