Add auth token to GET for /api/users/patient
Pass token from the three locations in the frontend where this route is called. Then authenticate the token in the backend and only return select data for all users if the user has an appropriate role.
Also refactor to use success rather than error and log error to console if it occurs.