Add auth token for /api/payments/initialise
Get auth token from currently signed in user. Pass to backend and verify there. Extract the users name and email from the token.
Can remove previous logic where name and email were taken from cookie and passed raw to the server. Also no need to validate a reCAPTCHA token on the server as this way supercedes that and is much more secure.