Fix payments webhook bug in dev mode
The payments webhook was erroring when a payment was made via the UI in development mode. This resulted in two Stripe customers being created due to the useEffect here being triggered twice. Since only one of the Stripe customer IDs got saved to Firestore, trying to find the patient associated with the other one resulted in not found and a crash.
Just wait a couple of seconds before running the useEffect to ensure all user updates have completed so it only runs once.