Fix bug so that name is saved to Stripe customer
We migrated from Firebase user to ID token when transferring data between client and server a week or two ago. However, this changed the name of the field from 'displayName' to 'name'. Now updated so the value is accessible and the name gets saved to the Stripe customer and is displayed in the Stripe dashboard again.