Remove attempted saving of phone number to auth
The phone number cannot be passed into the updateProfile method. It requires a different method but needs to be verified by the user first as the intention is for user authentication via phone.
We just wanted easy access to the number here, not to use phone as an auth method. Therefore remove it.