Fix issue with monthly subscription payments
Had previously just added 30 days to the current date for the pause duration. However, this would have created bugs such as not skipping a payment if the user paused immediately following a payment where the month had 31 days.
This updated solution uses the dayjs library as the functionality was not available on the native Date API. It aligns with the Stripe approach whereby payment is collected on the same day and time each month. If a pause were made today on 31st Jan it would resume on 29th Feb and not the beginning of March as the previous approach and Date would have.