Fix flaws in Firestore rules logic
The logic determining if the user owned the data was circular and always evaluated to true. Fixed this so that we can check if the uid from auth matches the one in the collection being mutated. This only works with the patients collection.
Disallow any mutations of the emails collection via Firestore rules. These will be permitted via the admin SDK instead.