Use admin SDK to set emails collection
Since we have prevented mutations by Firestore rules, we need to mutate using the admin SDK instead which only runs on the server and ignores the ruleset.
Note that the prior update to the patients collection will act as a check that the user is valid. If they fail here, the code for the admin SDK will never run.