Make no-unused-vars rule with ESLint less severe
Make the rule a warning rather than an error. This is better as it will still get flagged during linting or when running a build, but will not fail the build. Easier to develop this way since there is often temporary unused variables as a new approach is being tried.
Now we can uncomment the stripePriceIdLive in constants.js.