Fix issue with hero video sometimes not loading
There was a bug whereby the video in the hero section would not load when navigating from the dashboard. The page had to be manually refreshed for it to play.
This was because there was an extra route segment in the dashboard URL so the relative src could not find the video. Made the path absolute instead to resolve the issue.