Disable remote identity webhook when testing locally
Was having an issue whereby locally triggered events were being processed by both the local and remote webhooks. This caused writes to the production database while emulators were running.
Specifying from where the event was triggered in the metadata allows the event to be ignored when processed by the remote webhook if it was generated locally.