Disable remote payments webhook when testing locally
Was having an issue whereby locally triggered events were being processed by both the local and remote webhooks. This caused the remote webhook to fail while emulators were running as it was attempting to process a customer that didn't exist on the production database.
Now exits gracefully with a 204 response indicating what happened. This will be visible in the Stripe dashboard.