Change all 204 webhook responses to 202
During local CLI testing, invoking a 204 response was deemed invalid and causing an error. Stripe does not seem to allow in in the webhook, change to 202 instead which was fine.
Also remove a logging statement that never gets printed and will be accessible in the return JSON if the default case runs.