Fix issues with logging in locally during signup
Was encountering an issue whereby the user wouldn't login after account creation during the signup flow. This was because the emulators were only running on the server and not the client. The result was the client attempting to login using the emulated users credentials which did not exist in production.