Fix identity webhook failure to recognise environment
The webhook was indicating that the request was from a local source even when running from the remote hosted site.
Just checking if isOriginLocal is truthy is not good enough as it always evaluates to true. The values it can be assigned are the strings 'true' and 'false', not the boolean values due to constraints of the Stripe metadata object.