Fix emulator issue on create-account page
The request was erroring when trying to login after account creation. This was because of the attempt to start the emulator when it was already running.
The solution is to check the environment variables for the presence of auth emulator hostname which gets assigned on emulator start. If it's present, don't start the emulator.