Only create Stripe customer if one doesn't exist
Check to see if the Stripe customer ID has been saved to Firestore previously. Then only create a new customer if one didn't exist previously.
This is to future-proof the code for the introduction of other verticals. We will be reusing the same stripeUid for other products (but a different subscription).