Fix navbar transparency on the homepage
The removal of the /home route segment meant that the logic targeting navbar opacity on the homepage was no longer effective.
Rewrote it by targeting light pages instead of dark since the root (i.e. home) page couldn't be added to the darkPages array due to lack of corresponding route segment.