Use Admin SDK to create user from admin dashboard
Better to use the Admin SDK (firebase-admin/auth) to create a new user from the admin dashboard. This replaces the last implementation of the Client SDK for new user creation.
The client-side creation of accounts can now be disabled from the Firebase dashboard and this will prevent spam account creation.